Integrate Website's Watched List
Refactor ChangeWatched to avoid duplication, extend it to update the
website's watched list, and change its signature to require watched
parameter and conform to naming convention.
Add plugin menu option to browse and modify the website's watched and to
watch lists.
Add context menus to add items to watched and to watch list.
Add feature, enabled by a new setting, that adds movies from website's
watch list to the XBMC library along with regular subscription updates.
Modify create_item and build_listitem to accept the name of the calling
function as a parameter.  This allows special treatment for certain
menus.